Linux SPI speed cfg
On Linux devices the SPI speed can be set using the RF24 radio(ce,cs,spi_speed); constructor. This was not properly implemented.
- Defined RF24_SPI_SPEED for all devices in RF24_config.h
- Set spi_speed var in constructor
- Use spi_speed var for spi_transactions
- Default to RF24_SPI_SPEED when using the old BCM2835 enums
- Reconfig spi_speed var as uint32_t
- Update spi drivers for RPi,SPIDEV,wiringPi and MRAA

+++ RF24.h
@@ -79,7 +79,7 @@ private:
 
     uint16_t ce_pin; /**< "Chip Enable" pin, activates the RX or TX role */
     uint16_t csn_pin; /**< SPI Chip select */
-    uint16_t spi_speed; /**< SPI Bus Speed */
+    uint32_t spi_speed; /**< SPI Bus Speed */
     #if defined (RF24_LINUX) || defined (XMEGA_D3)
     uint8_t spi_rxbuff[32+1] ; //SPI receive buffer (payload max 32 bytes)
     uint8_t spi_txbuff[32+1] ; //SPI transmit buffer (payload max 32 bytes + 1 byte for the command)
